+++ morpheus-graphql-app/morpheus-graphql-app.cabal
@@ -135,6 +135,7 @@ library
       Data.Morpheus.App.Internal.Resolving.Batching
       Data.Morpheus.App.Internal.Resolving.Event
       Data.Morpheus.App.Internal.Resolving.MonadResolver
+      Data.Morpheus.App.Internal.Resolving.Refs
       Data.Morpheus.App.Internal.Resolving.Resolver
       Data.Morpheus.App.Internal.Resolving.ResolverState
       Data.Morpheus.App.Internal.Resolving.ResolveValue
